Add files via upload
The mechanism of the project is like any other online shopping website. There is one master page. The master page includes the category of shopping i.e vegetables and fruits accordingly along with the cart, about us, and payment option thereafter. A user has to first register and then log in. The user will be accessible to the content pages(which are not part of master page) only after logging in successfully. Once the user logged in, it is now feasible to select from multiple options of fruits and veggies. On selection of fruit and quantity of fruits and vegetables, he will be specified the bill on each new item added. Once the user selected items and is ready to proceed he will be directed to the ‘cart’ page to confirm his orders and accordingly the bill. Lastly to pay the bill the user will be directed to the ‘Payment’ page where his card details and address will be stored in the database. While building the project, all the requirements like master page, content page, skin files, state management techniques, database, calendar controls and validators are included.